feat: render all replies
