+++ controllers/tweet-controller.js
@@ -49,31 +49,31 @@ const tweetController = {
     } catch (err) {
       next(err)
     }
-  },
-  getTweet: (req, res, next) => {
-    const currentUser = helper.getUser(req)
-    const personalTweetRoute = true
-    const tweetId = req.params.id
-    Tweet.findByPk(tweetId, {
-      include: [
-        User,
-        { model: Like, as: 'LikedUsers' },
-        { model: Reply }
-      ]
-    })
-      .then(tweet => {
-        if (!tweet) throw new Error('No tweet found!')
-        const { dataValues, Replies, User, LikedUsers } = tweet
-        tweet = {
-          ...dataValues,
-          user: User.dataValues,
-          Replies,
-          repliesCount: Replies.length,
-          likesCount: LikedUsers.filter(likedUser => likedUser.isLike).length,
-          isLiked: LikedUsers.some(likedUser => likedUser.UserId === currentUser.id && likedUser.isLike)
-        }
-        res.render('tweet', { tweet, personalTweetRoute, currentUser })
-      })
   }
+  // getTweet: (req, res, next) => {
+  //   const currentUser = helper.getUser(req)
+  //   const personalTweetRoute = true
+  //   const tweetId = req.params.id
+  //   Tweet.findByPk(tweetId, {
+  //     include: [
+  //       User,
+  //       { model: Like, as: 'LikedUsers' },
+  //       { model: Reply }
+  //     ]
+  //   })
+  //     .then(tweet => {
+  //       if (!tweet) throw new Error('No tweet found!')
+  //       const { dataValues, Replies, User, LikedUsers } = tweet
+  //       tweet = {
+  //         ...dataValues,
+  //         user: User.dataValues,
+  //         Replies,
+  //         repliesCount: Replies.length,
+  //         likesCount: LikedUsers.filter(likedUser => likedUser.isLike).length,
+  //         isLiked: LikedUsers.some(likedUser => likedUser.UserId === currentUser.id && likedUser.isLike)
+  //       }
+  //       res.render('tweet', { tweet, personalTweetRoute, currentUser })
+  //     })
+  // }
 }
 module.exports = tweetController
